refactor(core): home the merged mode resolver in the core domain layer
The domain-layer consumers (core.speak, core.stop_snapshot) cannot import the
orchestration layer, so the unified resolver lives in teatree.core.mode_resolution
(reusing the preset override/schedule resolver, a domain leaf). Adds the tach edge
teatree.core -> teatree.loop.preset_resolution, mirroring the existing loop_cadences
edge. All migrated consumers + the resolver suite pass (161 tests).
