+++ src/teatree/core/management/commands/loops_tick.py
@@ -11,7 +11,7 @@ Each per-loop tick first reconciles the operating mode (#2544, #61): both driver
 that fire this command — the ``t3 worker``'s deadlined subprocess timer tick
 (``python -m teatree loops_tick --loop <name>``) and a manual by-hand
 ``t3 loops tick --loop <name>`` — converge here, so consulting
-:func:`teatree.loop.mode_resolution.resolve_active_mode` in ONE place reconciles both
+:func:`teatree.core.mode_resolution.resolve_active_mode` in ONE place reconciles both
 drivers identically. When the resolved mode's ``pauses_self_pump`` is true (a
 holiday-away mode only), the tick is skipped silently (parked) before any lease is
 claimed or overlay is preflighted; an autonomous-away mode defers questions but does
@@ -56,7 +56,7 @@ from teatree.core.backend_factory import code_host_from_overlay, iter_overlay_ba
 from teatree.core.loop_lease_manager import PER_LOOP_TICK_MUTEX_PREFIX, per_loop_owner_slot
 from teatree.core.models import LoopLease
 from teatree.loop.loop_cadences import loop_owner_ttl_seconds
-from teatree.loop.mode_resolution import resolve_active_mode
+from teatree.core.mode_resolution import resolve_active_mode
 from teatree.loop.preset_resolution import active_overlay_scope
 from teatree.loop.statusline import set_overridden_loops_reader, set_preset_line_reader
 from teatree.loops.preset_status import overridden_loop_names, preset_line_handles
