+++ src/teatree/core/mode_resolution.py
