+++ src/teatree/core/speak.py
@@ -65,6 +65,7 @@ from typing import IO
 from teatree.config import get_effective_settings
 from teatree.core import presence
 from teatree.core.backend_protocols import MessagingBackend
+from teatree.core.mode_resolution import resolve_active_mode
 from teatree.core.modelkit.notify_policy import NotifyAudience
 from teatree.core.speak_cleaning import clean_for_speech
 from teatree.paths import get_data_dir
@@ -138,8 +139,6 @@ def _is_away() -> bool:
     A resolution failure degrades to NOT away (returns ``False``): the
     away-gate must never suppress local audio on a transient error.
     """
-    from teatree.loop.mode_resolution import resolve_active_mode
-
     try:
         return resolve_active_mode().defers_questions
     except Exception as exc:  # noqa: BLE001 — a resolution failure must never mute local audio
