+++ src/teatree/dash/health_bands.py
@@ -25,7 +25,7 @@ from teatree.core.models.task_attempt import TaskAttempt
 from teatree.core.models.usage_window_state import UsageWindowState
 from teatree.core.selectors import build_headless_queue, build_interactive_queue
 from teatree.dash.gate_state import dash_gate_fail_open
-from teatree.loop.mode_resolution import resolve_active_mode
+from teatree.core.mode_resolution import resolve_active_mode
 from teatree.loops.live import LoopStatusReport, build_report
 
 logger = logging.getLogger(__name__)
