+++ src/teatree/dash/loop_control.py
@@ -17,7 +17,7 @@ from teatree.config import get_effective_settings
 from teatree.core.models.loop import Loop
 from teatree.core.models.loop_state import LoopState, LoopStatus
 from teatree.dash.gate_state import dash_gate_fail_open
-from teatree.loop.mode_resolution import resolve_active_mode
+from teatree.core.mode_resolution import resolve_active_mode
 from teatree.loops.preset_status import LoopVerdict, effective_verdicts
 
 logger = logging.getLogger(__name__)
