+++ tach.toml
@@ -175,7 +175,11 @@ depends_on = [
   "teatree.core.managers_overlay",
   "teatree.core.loop_lease_manager",
   "teatree.core.session_identity",
-  "teatree.loop.loop_cadences"
+  "teatree.loop.loop_cadences",
+  # The merged operating-mode resolver (teatree.core.mode_resolution, #61) reuses
+  # the preset override/schedule resolver — a domain leaf depending only on
+  # core.models, mirroring the loop_cadences edge above.
+  "teatree.loop.preset_resolution"
 ]
 
 [[modules]]
