+++ tests/teatree_dash/views/test_loops.py
@@ -9,7 +9,7 @@ from teatree.core.models.config_setting import ConfigSetting
 from teatree.core.models.loop import Loop
 from teatree.core.models.loop_state import LoopState, LoopStatus
 from teatree.dash.loop_control import AVAILABILITY_ACTIONS, LOOP_ACTIONS
-from teatree.loop.mode_resolution import clear_mode_override, resolve_active_mode
+from teatree.core.mode_resolution import clear_mode_override, resolve_active_mode
 
 
 def _make_loop(name: str = "dashloop") -> Loop:
